update code cmm013 (not fix)

+++ nts.uk/ctx/basic/nts.uk.ctx.basic.app/src/main/java/nts/uk/ctx/basic/app/command/organization/position/DeleteHistoryCommandHandler.java
@@ -33,38 +33,28 @@ public class DeleteHistoryCommandHandler extends CommandHandler<DeleteHistoryCom
 			
 			DeleteHistoryCommand command = context.getCommand();
 			String companyCode = AppContexts.user().companyCode();
-			//dudt
-//			if (positionRepository.ExistedHistory(command.getHistoryId())) {
-//				jobHistFinder.init();
-//			} else {
-//				JobHistory jobHistory = command.toDomain();
-//				jobHistory.validate();
-//				positionRepository.addHistory(jobHistory);
-//				GeneralDate pos1 = context.getCommand().getStartDate();
-//				GeneralDate endDate = pos1.addDays(-1);
-////				Optional<JobHistory> historyEndDate = positionRepository.getHistoryByEdate(companyCode, pos1);
-////				if (historyEndDate.isPresent()) {
-////					JobHistory jobHis = historyEndDate.get();
-////					jobHis.setEndDate(endDate);
-////					positionRepository.updateHistory(jobHis);
-//					positionRepository.deleteHist(companyCode, context.getCommand().getHistoryId());;
-	//
-//				}
-//			}
+
 			
 			String historyId = command.getHistoryId();
+			//check xem lich su co ton tai khong
+			repository.findSingleHistory(companyCode, historyId).ifPresent(x -> {throw new BusinessException("ER010");});
+			if(!repository.findSingleHistory(companyCode, historyId).isPresent()){
+				throw new BusinessException("ER010");
+			}
 			//xoa lich su
 			repository.deleteHist(companyCode, historyId);
 			//xoa tat ca postition cua lich su
 			repository.deleteJobTitleByHisId(companyCode, historyId);
 			//update tat ca ngay ket thuc cua lich su truoc ve 1999-12-31
 			//---Tim lich su cua thang dang truoc
-			GeneralDate endDate = GeneralDate.localDate(LocalDate.parse(command.getStartDate())).addDays(-1);
-			Optional<JobHistory> getHistoryByEndDate = repository.getHistoryByEdate(historyId, endDate);
-			JobHistory history = getHistoryByEndDate.get();
-			history.setEndDate(GeneralDate.localDate(LocalDate.parse("9999-12-31")));
-			//---update ngay ket thuc cua thang truoc
-			repository.updateHistory(history);
+			GeneralDate endDate = GeneralDate.localDate(LocalDate.parse(command.getOldStartDate())).addDays(-1);
+			Optional<JobHistory> getHistoryByEndDate = repository.getHistoryByEdate(companyCode, endDate);
+			if(getHistoryByEndDate.isPresent()){
+				JobHistory history = getHistoryByEndDate.get();
+				history.setEndDate(GeneralDate.localDate(LocalDate.parse("9999-12-31")));
+				//---update ngay ket thuc cua thang truoc
+				repository.updateHistory(history);
+			}
 			//Xoa tat ca quyen cua position
 					
 			//dudt
